feat: env var to select Py image for Petstore
Change documented in the Petstore README.md.

Also fixed a permission issue arising during Docker build, because the
Petstore docker-compose configuration creates the database directory
inside the Docker build context, which is not accessible as it is
created by user `root`. This affects the builds of both the FOCA and the
Petstore image and causes the Petstore app build to fail if the database
directory already exists (e.g., when rebuilding after the first build).
The `examples/` and `data/` directories were now added to
`.dockerignore` files for FOCA and Petstore builds respectively.

Finally, removed some extra whitespace in two files.